Package display card classes together
LocationCard, LocationListPanel, PersonCard, PersonListPanel,
VisitCard, VisitListPanel have been packaged together.
