+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -16,6 +16,9 @@ import seedu.address.logic.Logic;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.ui.list.LocationListPanel;
+import seedu.address.ui.list.PersonListPanel;
+import seedu.address.ui.list.VisitListPanel;
 
 /**
  * The Main Window. Provides the basic application layout containing
