+++ src/main/java/seedu/address/ui/list/LocationCard.java
@@ -1,17 +1,18 @@
-package seedu.address.ui;
+package seedu.address.ui.list;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.address.model.visit.Visit;
+import seedu.address.model.location.Location;
+import seedu.address.ui.UiPart;
 
 /**
  * An UI component that displays information of a {@code Person}.
  */
-public class VisitCard extends UiPart<Region> {
+public class LocationCard extends UiPart<Region> {
 
-    private static final String FXML = "VisitListCard.fxml";
+    private static final String FXML = "LocationListCard.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
@@ -21,29 +22,30 @@ public class VisitCard extends UiPart<Region> {
      * @see <a href="https://github.com/se-edu/addressbook-level4/issues/336">The issue on AddressBook level 4</a>
      */
 
-    public final Visit visit;
+    public final Location location;
 
     @FXML
     private HBox cardPane;
-    @FXML
-    private Label date;
-    @FXML
-    private Label id;
-    @FXML
-    private Label personName;
+
     @FXML
     private Label locationName;
+
+    @FXML
+    private Label address;
+
+    @FXML
+    private Label id;
+
 
     /**
      * Creates a {@code PersonCode} with the given {@code Person} and index to display.
      */
-    public VisitCard(Visit visit, int displayedIndex) {
+    public LocationCard(Location location, int displayedIndex) {
         super(FXML);
-        this.visit = visit;
+        this.location = location;
         id.setText(displayedIndex + ". ");
-        date.setText("Date: " + visit.getDate().toString());
-        personName.setText("Visit by: " + visit.getPersonName(visit.getPersonId()));
-        locationName.setText("Location: " + visit.getLocationName(visit.getLocationId()));
+        locationName.setText(location.getName().fullName);
+        address.setText(location.getAddress().value);
     }
 
     @Override
@@ -54,13 +56,13 @@ public class VisitCard extends UiPart<Region> {
         }
 
         // instanceof handles nulls
-        if (!(other instanceof PersonCard)) {
+        if (!(other instanceof LocationCard)) {
             return false;
         }
 
         // state check
-        VisitCard card = (VisitCard) other;
+        LocationCard card = (LocationCard) other;
         return id.getText().equals(card.id.getText())
-                && visit.equals(card.visit);
+                && location.equals(card.location);
     }
 }
